Add tests for detailsSlice
